Fix of the problem when the first row in Excel has an index=2 in case when the table doesn't contain any header
When the table doesn't contain any header then the first row should have index=1 (in Excel). But we cannot detect if the sheet is empty or there is only one row with index=0 with getLastRowNum function. So in this case we have to use getPhysicalNumberOfRows function and then we are able to detect if the sheet is empty or not.