CodeMirror: Do not HTML escape suggestions (#221)
I was using pe:codeMirror with the autocomplete feature via a bean method. The bean method returned several suggestions, including `&&`. The problem was that this got displayed as `&amp;&amp;` in the hint overlay from CodeMirror. The suggestions are send back to the client correctly, but when building the suggestions array, the PrimeFaces extensions CodeMirror widget uses `jQuery.html` instead of `jQuery.text`. This returns the text escaped for HTML. This suggestions array is then passed on as the `list` option of the [show hint addon](https://codemirror.net/doc/manual.html#addon_show-hint). This option expects an array of strings. The show hint addon seems to be handling HTML escaping properly, so we should not be doing it. Perhaps it was different in a previous version of CodeMirror?